Create the service test and fix some parameters types

+++ src/Console/MakeDrinkCommand.php
@@ -53,7 +53,7 @@ final class MakeDrinkCommand extends Command
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $drinkType = strtolower($input->getArgument('drink-type'));
-        $money = (int) $input->getArgument('money') ?? null;
+        $money = (float) $input->getArgument('money') ?? null;
         $sugars = (int) $input->getArgument('sugars') ?? null;
         $extraHot = (bool) $input->getOption('extra-hot');
